refactured main.rs in a lot of files to better maintain
